use stream information from index ranges when present
this moves the index range handling from into the Searches class and pushes the filter handling down, to make invoking the methods easier.
index ranges now match when the times match, for stream searches the current index set must be the target of the stream, or an index range must contain explicit information about which streams are contained in the index (which is performed by index range calculation now)
