+++ graylog2-server/src/main/java/org/graylog2/indexer/IndexHelper.java
@@ -16,15 +16,11 @@
  */
 package org.graylog2.indexer;
 
-import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.ImmutableSortedSet;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
 
 import org.elasticsearch.index.query.QueryBuilder;
 import org.elasticsearch.index.query.QueryBuilders;
-import org.graylog2.indexer.ranges.IndexRange;
-import org.graylog2.indexer.ranges.IndexRangeService;
 import org.graylog2.plugin.Tools;
 import org.graylog2.plugin.indexer.searches.timeranges.TimeRange;
 
@@ -73,30 +69,4 @@ public class IndexHelper {
         return r;
     }
 
-    public static Set<String> determineAffectedIndices(IndexRangeService indexRangeService,
-                                                       TimeRange range,
-                                                       IndexSet indexSet) {
-        final Set<IndexRange> indexRanges = determineAffectedIndicesWithRanges(indexRangeService, range, indexSet);
-        final ImmutableSet.Builder<String> indices = ImmutableSet.builder();
-        for (IndexRange indexRange : indexRanges) {
-            indices.add(indexRange.indexName());
-        }
-
-        return indices.build();
-    }
-
-    public static Set<IndexRange> determineAffectedIndicesWithRanges(IndexRangeService indexRangeService,
-                                                                     TimeRange range,
-                                                                     IndexSet indexSet) {
-        final ImmutableSortedSet.Builder<IndexRange> indices = ImmutableSortedSet.orderedBy(IndexRange.COMPARATOR);
-        for (IndexRange indexRange : indexRangeService.find(range.getFrom(), range.getTo())) {
-            // if we only consider a certain index set, filter the index ranges by it.
-            if (indexSet != null && !indexSet.isManagedIndex(indexRange.indexName())) {
-                continue;
-            }
-            indices.add(indexRange);
-        }
-
-        return indices.build();
-    }
 }
