+++ graylog2-server/src/main/java/org/graylog2/indexer/searches/Searches.java
@@ -16,6 +16,7 @@
  */
 package org.graylog2.indexer.searches;
 
+import com.google.common.collect.ImmutableSortedSet;
 import com.google.common.collect.Sets;
 
 import com.codahale.metrics.Histogram;
@@ -67,9 +68,11 @@ import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
+import java.util.SortedSet;
 import java.util.concurrent.TimeUnit;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import java.util.stream.Collectors;
 
 import javax.annotation.Nullable;
 import javax.inject.Inject;
@@ -96,7 +99,6 @@ public class Searches {
     public static final String AGG_CARDINALITY = "gl2_field_cardinality";
     public static final String AGG_VALUE_COUNT = "gl2_value_count";
     private static final Pattern filterStreamIdPattern = Pattern.compile("^(.+[^\\p{Alnum}])?streams:([\\p{XDigit}]+)");
-
 
     public enum TermsStatsOrder {
         TERM,
@@ -180,8 +182,7 @@ public class Searches {
     }
 
     public CountResult count(String query, TimeRange range, String filter) {
-        IndexSet indexSet = getIndexSetForFilter(filter);
-        Set<String> indices = IndexHelper.determineAffectedIndices(indexRangeService, range, indexSet);
+        Set<String> indices = determineAffectedIndices(range, filter);
 
         final SearchRequestBuilder srb;
         if (filter == null) {
@@ -197,8 +198,7 @@ public class Searches {
     }
 
     public ScrollResult scroll(String query, TimeRange range, int limit, int offset, List<String> fields, String filter) {
-        IndexSet indexSet = getIndexSetForFilter(filter);
-        final Set<String> indices = IndexHelper.determineAffectedIndices(indexRangeService, range, indexSet);
+        final Set<String> indices = determineAffectedIndices(range, filter);
 
         // only request the fields we asked for otherwise we can't figure out which fields will be in the result set
         // until we've scrolled through the entire set.
@@ -245,8 +245,7 @@ public class Searches {
     }
 
     public SearchResult search(SearchesConfig config) {
-        IndexSet indexSet = getIndexSetForFilter(config.filter());
-        Set<IndexRange> indices = IndexHelper.determineAffectedIndicesWithRanges(indexRangeService, config.range(), indexSet);
+        Set<IndexRange> indices = determineAffectedIndicesWithRanges(config.range(), config.filter());
 
         Set<String> indexNames = Sets.newHashSet();
         for (IndexRange index : indices) {
@@ -261,31 +260,16 @@ public class Searches {
         return new SearchResult(r.getHits(), indices, config.query(), request.source(), r.getTook());
     }
 
-    @Nullable
-    private IndexSet getIndexSetForFilter(String filter) {
-        final Optional<String> streamId = extractStreamId(filter);
-        IndexSet indexSet = null;
-        if (streamId.isPresent()) {
-            try {
-                final Stream stream = streamService.load(streamId.get());
-                indexSet = stream.getIndexSet();
-            } catch (NotFoundException ignored) {
-            }
-        }
-        return indexSet;
-    }
-
     public TermsResult terms(String field, int size, String query, String filter, TimeRange range) {
         if (size == 0) {
             size = 50;
         }
 
-        IndexSet indexSet = getIndexSetForFilter(filter);
         SearchRequestBuilder srb;
         if (filter == null) {
-            srb = standardSearchRequest(query, IndexHelper.determineAffectedIndices(indexRangeService, range, indexSet), range);
+            srb = standardSearchRequest(query, determineAffectedIndices(range, null), range);
         } else {
-            srb = filteredSearchRequest(query, filter, IndexHelper.determineAffectedIndices(indexRangeService, range, indexSet), range);
+            srb = filteredSearchRequest(query, filter, determineAffectedIndices(range, filter), range);
         }
 
         FilterAggregationBuilder builder = AggregationBuilders.filter(AGG_FILTER)
@@ -325,12 +309,11 @@ public class Searches {
             size = 50;
         }
 
-        IndexSet indexSet = getIndexSetForFilter(filter);
         SearchRequestBuilder srb;
         if (filter == null) {
-            srb = standardSearchRequest(query, IndexHelper.determineAffectedIndices(indexRangeService, range, indexSet), range);
+            srb = standardSearchRequest(query, determineAffectedIndices(range, filter), range);
         } else {
-            srb = filteredSearchRequest(query, filter, IndexHelper.determineAffectedIndices(indexRangeService, range, indexSet), range);
+            srb = filteredSearchRequest(query, filter, determineAffectedIndices(range, filter), range);
         }
 
 
@@ -421,12 +404,11 @@ public class Searches {
                                        boolean includeStats,
                                        boolean includeCount)
             throws FieldTypeException {
-        IndexSet indexSet = getIndexSetForFilter(filter);
         SearchRequestBuilder srb;
         if (filter == null) {
-            srb = standardSearchRequest(query, IndexHelper.determineAffectedIndices(indexRangeService, range, indexSet), range);
+            srb = standardSearchRequest(query, determineAffectedIndices(range, filter), range);
         } else {
-            srb = filteredSearchRequest(query, filter, IndexHelper.determineAffectedIndices(indexRangeService, range, indexSet), range);
+            srb = filteredSearchRequest(query, filter, determineAffectedIndices(range, filter), range);
         }
 
         FilterAggregationBuilder builder = AggregationBuilders.filter(AGG_FILTER)
@@ -480,8 +462,7 @@ public class Searches {
         QueryStringQueryBuilder qs = queryStringQuery(query);
         qs.allowLeadingWildcard(configuration.isAllowLeadingWildcardSearches());
 
-        IndexSet indexSet = getIndexSetForFilter(filter);
-        final Set<String> affectedIndices = IndexHelper.determineAffectedIndices(indexRangeService, range, indexSet);
+        final Set<String> affectedIndices = determineAffectedIndices(range, filter);
         final SearchRequestBuilder srb = c.prepareSearch(affectedIndices.toArray(new String[affectedIndices.size()]))
                 .setIndicesOptions(IndicesOptions.lenientExpandOpen())
                 .setQuery(qs)
@@ -522,9 +503,8 @@ public class Searches {
         QueryStringQueryBuilder qs = queryStringQuery(query);
         qs.allowLeadingWildcard(configuration.isAllowLeadingWildcardSearches());
 
-        IndexSet indexSet = getIndexSetForFilter(filter);
         SearchRequestBuilder srb = c.prepareSearch();
-        final Set<String> affectedIndices = IndexHelper.determineAffectedIndices(indexRangeService, range, indexSet);
+        final Set<String> affectedIndices = determineAffectedIndices(range, filter);
         srb.setIndices(affectedIndices.toArray(new String[affectedIndices.size()]));
         srb.setQuery(qs);
         srb.addAggregation(builder);
@@ -697,7 +677,7 @@ public class Searches {
      * This is currently a workaround. A better solution would be to pass the stream id which is supposed to be the scope
      * for a search query as a separate parameter.
      *
-     * @param filter
+     * @param filter the filter string like "streams:xxxyyyzzz"
      * @return the optional stream id
      */
     public static Optional<String> extractStreamId(String filter) {
@@ -710,4 +690,54 @@ public class Searches {
         }
         return Optional.empty();
     }
+
+    public Set<String> determineAffectedIndices(TimeRange range,
+                                                @Nullable String filter) {
+        final Set<IndexRange> indexRanges = determineAffectedIndicesWithRanges(range, filter);
+        return indexRanges.stream()
+                .map(IndexRange::indexName)
+                .collect(Collectors.toSet());
+    }
+
+    public Set<IndexRange> determineAffectedIndicesWithRanges(TimeRange range,
+                                                              @Nullable String filter) {
+        final Optional<String> streamId = extractStreamId(filter);
+        IndexSet indexSet = null;
+        // if we are searching in a stream, we are further restricting the indices using the currently
+        // configure index set of that stream.
+        // later on we will also test against each index range (we load all of them) to see if there are
+        // additional index ranges that match, this can happen with restored archives or when the index set for
+        // a stream has changed: a stream only knows about its currently configured index set, no the history
+        if (streamId.isPresent()) {
+            try {
+                final Stream stream = streamService.load(streamId.get());
+                indexSet = stream.getIndexSet();
+            } catch (NotFoundException ignored) {
+            }
+        }
+
+        final ImmutableSortedSet.Builder<IndexRange> indices = ImmutableSortedSet.orderedBy(IndexRange.COMPARATOR);
+        final SortedSet<IndexRange> indexRanges = indexRangeService.find(range.getFrom(), range.getTo());
+        for (IndexRange indexRange : indexRanges) {
+            // if we aren't in a stream search, we look at all the ranges matching the time range.
+            if (indexSet == null && filter == null) {
+                indices.add(indexRange);
+                continue;
+            }
+
+            // A range applies to this search if either: the current index set of the stream matches or a previous index set matched.
+            final boolean streamInIndexRange = streamId.isPresent() && indexRange.streamIds() != null && indexRange.streamIds().contains(streamId.get());
+            final boolean streamInCurrentIndexSet = indexSet != null && indexSet.isManagedIndex(indexRange.indexName());
+
+            if (streamInIndexRange) {
+                indices.add(indexRange);
+            }
+            if (streamInCurrentIndexSet) {
+                indices.add(indexRange);
+            }
+        }
+
+        return indices.build();
+    }
+
 }
