+++ graylog2-server/src/test/java/org/graylog2/indexer/IndexHelperTest.java
@@ -17,17 +17,8 @@
 package org.graylog2.indexer;
 
 import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.ImmutableSortedSet;
 
-import org.elasticsearch.index.query.RangeQueryBuilder;
-import org.graylog2.indexer.ranges.IndexRange;
 import org.graylog2.indexer.ranges.IndexRangeService;
-import org.graylog2.indexer.ranges.MongoIndexRange;
-import org.graylog2.plugin.Tools;
-import org.graylog2.plugin.indexer.searches.timeranges.AbsoluteRange;
-import org.graylog2.plugin.indexer.searches.timeranges.KeywordRange;
-import org.graylog2.plugin.indexer.searches.timeranges.RelativeRange;
-import org.graylog2.plugin.indexer.searches.timeranges.TimeRange;
 import org.joda.time.DateTime;
 import org.joda.time.DateTimeUtils;
 import org.joda.time.DateTimeZone;
@@ -41,12 +32,9 @@ import org.mockito.junit.MockitoRule;
 
 import java.util.Map;
 import java.util.Optional;
-import java.util.SortedSet;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.ArgumentMatchers.startsWith;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
@@ -109,147 +97,5 @@ public class IndexHelperTest {
         final IndexSet indexSet = mock(IndexSet.class);
         when(indexSet.getManagedIndicesNames()).thenReturn(new String[0]);
         assertThat(IndexHelper.getOldestIndices(indexSet, 9001)).isEmpty();
-    }
-
-    @Test
-    public void determineAffectedIndicesWithRangesIncludesDeflectorTarget() throws Exception {
-        final DateTime now = DateTime.now(DateTimeZone.UTC);
-        final MongoIndexRange indexRange0 = MongoIndexRange.create("graylog_0", now, now.plusDays(1), now, 0);
-        final MongoIndexRange indexRange1 = MongoIndexRange.create("graylog_1", now.plusDays(1), now.plusDays(2), now, 0);
-        final MongoIndexRange indexRangeLatest = MongoIndexRange.create("graylog_2", new DateTime(0L, DateTimeZone.UTC), new DateTime(0L, DateTimeZone.UTC), now, 0);
-        final SortedSet<IndexRange> indices = ImmutableSortedSet.orderedBy(IndexRange.COMPARATOR)
-            .add(indexRange0)
-            .add(indexRange1)
-            .add(indexRangeLatest)
-            .build();
-
-        when(indexRangeService.find(any(DateTime.class), any(DateTime.class))).thenReturn(indices);
-
-        final TimeRange absoluteRange = AbsoluteRange.create(now.minusDays(1), now.plusDays(1));
-        final TimeRange keywordRange = KeywordRange.create("1 day ago");
-        final TimeRange relativeRange = RelativeRange.create(3600);
-
-        assertThat(IndexHelper.determineAffectedIndicesWithRanges(indexRangeService, absoluteRange, null))
-            .containsExactly(indexRangeLatest, indexRange0, indexRange1);
-        assertThat(IndexHelper.determineAffectedIndicesWithRanges(indexRangeService, keywordRange, null))
-            .containsExactly(indexRangeLatest, indexRange0, indexRange1);
-        assertThat(IndexHelper.determineAffectedIndicesWithRanges(indexRangeService, relativeRange, null))
-            .containsExactly(indexRangeLatest, indexRange0, indexRange1);
-    }
-
-    @Test
-    public void determineAffectedIndicesWithRangesDoesNotIncludesDeflectorTargetIfMissing() throws Exception {
-        final DateTime now = DateTime.now(DateTimeZone.UTC);
-        final MongoIndexRange indexRange0 = MongoIndexRange.create("graylog_0", now, now.plusDays(1), now, 0);
-        final MongoIndexRange indexRange1 = MongoIndexRange.create("graylog_1", now.plusDays(1), now.plusDays(2), now, 0);
-        final SortedSet<IndexRange> indices = ImmutableSortedSet.orderedBy(IndexRange.COMPARATOR)
-            .add(indexRange0)
-            .add(indexRange1)
-            .build();
-
-        when(indexRangeService.find(any(DateTime.class), any(DateTime.class))).thenReturn(indices);
-
-        final TimeRange absoluteRange = AbsoluteRange.create(now.minusDays(1), now.plusDays(1));
-        final TimeRange keywordRange = KeywordRange.create("1 day ago");
-        final TimeRange relativeRange = RelativeRange.create(3600);
-
-        assertThat(IndexHelper.determineAffectedIndicesWithRanges(indexRangeService, absoluteRange, null))
-            .containsExactly(indexRange0, indexRange1);
-        assertThat(IndexHelper.determineAffectedIndicesWithRanges(indexRangeService, keywordRange, null))
-            .containsExactly(indexRange0, indexRange1);
-        assertThat(IndexHelper.determineAffectedIndicesWithRanges(indexRangeService, relativeRange, null))
-            .containsExactly(indexRange0, indexRange1);
-    }
-
-    @Test
-    public void determineAffectedIndicesIncludesDeflectorTarget() throws Exception {
-        final DateTime now = DateTime.now(DateTimeZone.UTC);
-        final MongoIndexRange indexRange0 = MongoIndexRange.create("graylog_0", now, now.plusDays(1), now, 0);
-        final MongoIndexRange indexRange1 = MongoIndexRange.create("graylog_1", now.plusDays(1), now.plusDays(2), now, 0);
-        final MongoIndexRange indexRangeLatest = MongoIndexRange.create("graylog_2", new DateTime(0L, DateTimeZone.UTC), new DateTime(0L, DateTimeZone.UTC), now, 0);
-        final SortedSet<IndexRange> indices = ImmutableSortedSet.orderedBy(IndexRange.COMPARATOR)
-            .add(indexRange0)
-            .add(indexRange1)
-            .add(indexRangeLatest)
-            .build();
-
-        when(indexRangeService.find(any(DateTime.class), any(DateTime.class))).thenReturn(indices);
-
-        final TimeRange absoluteRange = AbsoluteRange.create(now.minusDays(1), now.plusDays(1));
-        final TimeRange keywordRange = KeywordRange.create("1 day ago");
-        final TimeRange relativeRange = RelativeRange.create(3600);
-
-        assertThat(IndexHelper.determineAffectedIndices(indexRangeService, absoluteRange, null))
-            .containsExactly(indexRangeLatest.indexName(), indexRange0.indexName(), indexRange1.indexName());
-        assertThat(IndexHelper.determineAffectedIndices(indexRangeService, keywordRange, null))
-            .containsExactly(indexRangeLatest.indexName(), indexRange0.indexName(), indexRange1.indexName());
-        assertThat(IndexHelper.determineAffectedIndices(indexRangeService, relativeRange, null))
-            .containsExactly(indexRangeLatest.indexName(), indexRange0.indexName(), indexRange1.indexName());
-    }
-
-    @Test
-    public void determineAffectedIndicesDoesNotIncludesDeflectorTargetIfMissing() throws Exception {
-        final DateTime now = DateTime.now(DateTimeZone.UTC);
-        final MongoIndexRange indexRange0 = MongoIndexRange.create("graylog_0", now, now.plusDays(1), now, 0);
-        final MongoIndexRange indexRange1 = MongoIndexRange.create("graylog_1", now.plusDays(1), now.plusDays(2), now, 0);
-        final SortedSet<IndexRange> indices = ImmutableSortedSet.orderedBy(IndexRange.COMPARATOR)
-            .add(indexRange0)
-            .add(indexRange1)
-            .build();
-
-        when(indexRangeService.find(any(DateTime.class), any(DateTime.class))).thenReturn(indices);
-
-        final TimeRange absoluteRange = AbsoluteRange.create(now.minusDays(1), now.plusDays(1));
-        final TimeRange keywordRange = KeywordRange.create("1 day ago");
-        final TimeRange relativeRange = RelativeRange.create(3600);
-
-        assertThat(IndexHelper.determineAffectedIndices(indexRangeService, absoluteRange, null))
-            .containsOnly(indexRange0.indexName(), indexRange1.indexName());
-        assertThat(IndexHelper.determineAffectedIndices(indexRangeService, keywordRange, null))
-            .containsOnly(indexRange0.indexName(), indexRange1.indexName());
-        assertThat(IndexHelper.determineAffectedIndices(indexRangeService, relativeRange, null))
-            .containsOnly(indexRange0.indexName(), indexRange1.indexName());
-    }
-
-    @Test
-    public void getTimestampRangeFilterReturnsNullIfTimeRangeIsNull() {
-        assertThat(IndexHelper.getTimestampRangeFilter(null)).isNull();
-    }
-
-    @Test
-    public void getTimestampRangeFilterReturnsRangeQueryWithGivenTimeRange() {
-        final DateTime from = new DateTime(2016, 1, 15, 12, 0, DateTimeZone.UTC);
-        final DateTime to = from.plusHours(1);
-        final TimeRange timeRange = AbsoluteRange.create(from, to);
-        final RangeQueryBuilder queryBuilder = (RangeQueryBuilder) IndexHelper.getTimestampRangeFilter(timeRange);
-        assertThat(queryBuilder)
-            .isNotNull()
-            .hasFieldOrPropertyWithValue("name", "timestamp")
-            .hasFieldOrPropertyWithValue("from", Tools.buildElasticSearchTimeFormat(from))
-            .hasFieldOrPropertyWithValue("to", Tools.buildElasticSearchTimeFormat(to));
-    }
-
-    @Test
-    public void determineAffectedIndicesFilterIndexPrefix() throws Exception {
-        final DateTime now = DateTime.now(DateTimeZone.UTC);
-        final MongoIndexRange indexRange0 = MongoIndexRange.create("graylog_0", now, now.plusDays(1), now, 0);
-        final MongoIndexRange indexRange1 = MongoIndexRange.create("graylog_1", now.plusDays(1), now.plusDays(2), now, 0);
-        final MongoIndexRange b0 = MongoIndexRange.create("b_0", now.plusDays(1), now.plusDays(2), now, 0);
-        final MongoIndexRange b1 = MongoIndexRange.create("b_1", now.plusDays(1), now.plusDays(2), now, 0);
-        final SortedSet<IndexRange> indices = ImmutableSortedSet.orderedBy(IndexRange.COMPARATOR)
-                .add(indexRange0)
-                .add(indexRange1)
-                .add(b0)
-                .add(b1)
-                .build();
-
-        when(indexRangeService.find(any(DateTime.class), any(DateTime.class))).thenReturn(indices);
-
-        final TimeRange absoluteRange = AbsoluteRange.create(now.minusDays(1), now.plusDays(1));
-        final IndexSet indexSet = mock(IndexSet.class);
-        when(indexSet.isManagedIndex(startsWith("b_"))).thenReturn(true);
-
-        assertThat(IndexHelper.determineAffectedIndices(indexRangeService, absoluteRange, indexSet))
-                .containsOnly(b0.indexName(), b1.indexName());
     }
 }
